// src/api/generated/models/userMeResponse.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * BFF Service
 * Backend For Frontend service aggregating microservices
 * OpenAPI spec version: 1.0.0
 */
import type { UserMeResponseUserType } from "./userMeResponseUserType";
import type { UserMeResponseUserDetails } from "./userMeResponseUserDetails";

/**
 * Response model for /me endpoint with complete user information.
 */
export interface UserMeResponse {
  user_id: string;
  name: string;
  email: string;
  groups?: string[];
  user_type?: UserMeResponseUserType;
  user_details?: UserMeResponseUserDetails;
}